+++ packages/terra-date-time-picker/src/terra-dev-site/test/date-time-picker/DateTimePickerOnChange.test.jsx
@@ -75,6 +75,7 @@ class DatePickerOnChange extends React.Component {
         <DateTimePicker
           name="date-time-picker-onchange"
           onChange={this.handleDateChange}
+          initialTimeZone="America/Chicago"
         />
       </div>
     );