+++ packages/terra-embedded-content-consumer/src/terra-dev-site/test/embedded-content-consumer/consumers/BasicConsumer.test.jsx
@@ -11,6 +11,7 @@ Consumer.init();
 const BasicConsumer = () => (
   <EmbeddedContentConsumer
     className={cx('iframe')}
+    options={{ iframeAttrs: { title: 'Basic Consumer' } }}
     src="/raw/provider/terra-embedded-content-consumer/embedded-content-consumer/providers/basic-provider"
   />
 );