+++ packages/terra-navigation-layout/tests/wdio/navigation-layout-spec.js
@@ -8,19 +8,19 @@ Terra.describeViewports('Navigation Layout', ['tiny', 'small', 'medium', 'large'
 
   it('Displays a complex NavigationLayout - page 1', () => {
     browser.url('/raw/tests/terra-navigation-layout/navigation-layout/navigation-layout-complex');
-    browser.click('#test-root .page-1-link');
+    $('#test-root .page-1-link').click();
     Terra.validates.element('complex - page 1', { selector });
   });
 
   it('Displays a complex NavigationLayout - page 2', () => {
     browser.url('/raw/tests/terra-navigation-layout/navigation-layout/navigation-layout-complex');
-    browser.click('#test-root .page-2-link');
+    $('#test-root .page-2-link').click();
     Terra.validates.element('complex - page 2', { selector });
   });
 
   it('Displays a complex NavigationLayout - page 3', () => {
     browser.url('/raw/tests/terra-navigation-layout/navigation-layout/navigation-layout-complex');
-    browser.click('#test-root .page-3-link');
+    $('#test-root .page-3-link').click();
     Terra.validates.element('complex - page 3', { selector });
   });
 });