+++ packages/terra-popup/src/terra-dev-site/test/popup/ArrowVerticalLeftAdjustmentPopup.test.jsx
@@ -74,6 +74,7 @@ class AlignmentPopup extends React.Component {
           className={cx('popup-button')}
           onClick={this.handleButtonClick}
           ref={this.setButtonNode}
+          aria-label="Alignment button"
         >
           {''}
         </button>