+++ packages/terra-application-utility/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated to terra-functional-testing
+
 ## 2.41.0 - (January 5, 2021)
 
 * Changed
@@ -9,7 +12,7 @@
   * Removed `intlShape` imports
 
 * Fixed
- * Fixed broken links in documentation.
+  * Fixed broken links in documentation.
 
 ## 2.40.0 - (November 24, 2020)