+++ packages/terra-application-utility/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.42.0 - (March 9, 2021)
+
 * Changed
   * Updated to terra-functional-testing